refactor: change dir for helper
